Update dependencies and enhance user management features; upgrade Next.js version, implement role-based access for user actions, and improve API integration for roles and entities. Refactor sidebar to include new permissions and streamline user form handling.

+++ app/api/users/[id]/route.js
@@ -28,8 +28,18 @@ export async function PUT(request, { params }) {
 }
 
 export async function DELETE(request, { params }) {
+  const userRole = await getUserById(id);
+
+  if (userRole.role !== "super admin") {
+    return NextResponse.json(
+      { error: "You are not authorized to delete this user" },
+      { status: 403 }
+    );
+  }
+
   try {
     const { id } = await params;
+
     const deletedProfile = await deleteProfile(id);
 
     if (!deletedProfile.success) {
